requirements.txt
# Core dependencies
anthropic>=0.25.0
chromadb>=0.4.22
sentence-transformers>=2.3.1
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Web interface
streamlit>=1.30.0

# Code quality
black>=24.0.0
ruff>=0.4.0

# Dependency management
pipreqs>=0.5.0
pip-audit>=2.7.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
